display: carve encode/pool worker.rs (pure move)
